// src/types.ts
export type ShortManifestResource =
  | "catalog"
  | "meta"
  | "stream"
  | "subtitles"
  | "addon_catalog";
export type Extra = "search" | "genre" | "skip";
export type ContentType = "movie" | "series" | "channel" | "tv";

export type DefaultConfig = Record<string, any> | undefined;
// ...
export interface Manifest {
  /**
   * Identifier, dot-separated, e.g. "com.stremio.filmon"
   */
  id: string;
  /**
   * Human readable name
   */
  name: string;
  /**
   *  Human readable description
   */
  description: string;
  /**
   * Semantic version of the addon
   */
  version: string;
  /**
   * Supported resources, defined as an array of objects (long version) or strings (short version).
   *
   * Example #1: [{"name": "stream", "types": ["movie"], "idPrefixes": ["tt"]}]
   *
   * Example #2: ["catalog", "meta", "stream", "subtitles", "addon_catalog"]
   */
  resources: Array<ShortManifestResource | FullManifestResource>;
  /**
   * Supported types.
   */
  types: string[];
  /**
   * Use this if you want your addon to be called only for specific content IDs.
   *
   * For example, if you set this to ["yt_id:", "tt"], your addon will only be called for id values that start with 'yt_id:' or 'tt'.
   */
  idPrefixes?: string[];
  /**
   * A list of the content catalogs your addon provides.
   *
   * Leave this an empty array ([]) if your addon does not provide the catalog resource.
   */
  catalogs: ManifestCatalog[];
  /**
   * Array of Catalog objects, a list of other addon manifests.
   *
   * This can be used for an addon to act just as a catalog of other addons.
   */
  addonCatalogs?: ManifestCatalog[];

  /**
   * A list of settings that users can set for your addon.
   */
  config?: ManifestConfig[];

  /**
   * Background image for the addon.
   *
   * URL to png/jpg, at least 1024x786 resolution.
   */
  background?: string;

  /**
   * @deprecated use `logo` instead.
   */
  icon?: string;

  /**
   * Logo icon, URL to png, monochrome, 256x256.
   */
  logo?: string;
  /**
   * Contact email for addon issues.
   * Used for the Report button in the app.
   * Also, the Stremio team may reach you on this email for anything relating your addon.
   */
  contactEmail?: string;
  behaviorHints?:
    | {
        /**
         * If the addon includes adult content.
         *
         * Defaults to false.
         */
        adult?: boolean;
        /**
         * If the addon includes P2P content, such as BitTorrent, which may reveal the user's IP to other streaming parties.
         *
         * Used to provide an adequate warning to the user.
         */
        p2p?: boolean;

        /**
         * Default is `false`. If the addon supports settings, it will add a button next to "Install" in Stremio that will point to the `/configure` path on the addon's domain. For more information, read [User Data](https://github.com/Stremio/stremio-addon-sdk/blob/master/docs/api/responses/manifest.md#user-data) (or if you are not using the Addon SDK, read: [Advanced User Data](https://github.com/Stremio/stremio-addon-sdk/blob/master/docs/advanced.md#using-user-data-in-addons) and [Creating Addon Configuration Pages](https://github.com/Stremio/stremio-addon-sdk/blob/master/docs/advanced.md#creating-addon-configuration-pages))
         */
        configurable?: boolean;

        /**
         * Default is `false`. If set to `true`, the "Install" button will not show for your addon in Stremio. Instead a "Configure" button will show pointing to the `/configure` path on the addon's domain. For more information, read [User Data](https://github.com/Stremio/stremio-addon-sdk/blob/master/docs/api/responses/manifest.md#user-data) (or if you are not using the Addon SDK, read: [Advanced User Data](https://github.com/Stremio/stremio-addon-sdk/blob/master/docs/advanced.md#using-user-data-in-addons) and [Creating Addon Configuration Pages](https://github.com/Stremio/stremio-addon-sdk/blob/master/docs/advanced.md#creating-addon-configuration-pages))
         */
        configurationRequired?: boolean;
      }
    | undefined;

  stremioAddonsConfig?: StremioAddonsConfig;
}

export type ManifestConfigType =
  | "text"
  | "number"
  | "password"
  | "checkbox"
  | "select";

/**
 * Addon setting.
 */
export interface ManifestConfig {
  /**
   * A key that will identify the user chosen value.
   */
  key: string;

  /**
   * The type of data that the setting stores.
   */
  type: ManifestConfigType;

  /**
   * The default value. For `type: "boolean"` this can be set to "checked" to default to enabled.
   */
  default?: string;

  /**
   * The title of the setting.
   */
  title?: string;

  /**
   * List of (string) choices for `type: "select"`
   */
  options?: string[];

  /**
   * If the value is required or not. Only applies to the following types: "string", "number". (default is `false`)
   */
  required?: boolean;
}

/**
 * Used as a response for defineResourceHandler.
 */
export interface AddonCatalog {
  /**
   * only http is currently officially supported.
   */
  transportName: string;
  /**
   * The URL of the addon's manifest.json file.
   */
  transportUrl: string;
  /**
   * Object representing the addon's Manifest Object.
   */
  manifest: Manifest;
}

export interface FullManifestResource {
  /**
   * Resource name.
   */
  name: ShortManifestResource;
  /**
   * Supported types.
   */
  types: string[];
  /**
   * Use this if you want your addon to be called only for specific content IDs
   *
   * For example, if you set this to ["yt_id:", "tt"], your addon will only be called for id values that start with 'yt_id:' or 'tt'.
   */
  idPrefixes?: string[];
}

export interface ManifestCatalog {
  /**
   *  This is the content type of the catalog.
   */
  type: string;
  /**
   * The id of the catalog, can be any unique string describing the catalog (unique per addon, as an addon can have many catalogs).
   *
   * For example: if the catalog name is "Favourite Youtube Videos", the id can be "fav_youtube_videos".
   */
  id: string;
  /**
   * Human readable name of the catalog.
   */
  name: string;
  /**
   * Use the 'options' property of 'extra' instead.
   * @deprecated
   */
  genres?: string[];
  /**
   * All extra properties related to this catalog.
   */
  extra?: ManifestExtra[];
}

export interface ManifestExtra {
  /**
   * The name of the property
   *
   * This name will be used in the extraProps argument itself.
   *
   * @example "search", "genre", "skip"
   */
  name: string;
  /**
   * Set to true if this property must always be passed.
   */
  isRequired?: boolean;
  /**
   * Possible values for this property.
   * This is useful for things like genres, where you need the user to select from a pre-set list of options.
   *
   * e.g. { name: "genre", options: ["Action", "Comedy", "Drama"] }
   *
   * It's also useful if we want to specify a limited number of pages (for the skip parameter).
   *
   * e.g. { name: "skip", options: ["0", "100", "200"] }
   */
  options?: string[];
  /**
   * The limit of values a user may select from the pre-set options list
   *
   * By default this is set to 1.
   */
  optionsLimit?: number;
}

export interface StremioAddonsConfig {
  issuer: "https://stremio-addons.net";
  signature: string;
}

/**
 * The addonInterface, as returned from builder.getInterface()
 */
export interface AddonInterface {
  manifest: Manifest;
  get: (
    resource: ShortManifestResource,
    type: ContentType,
    id: string,
    extra?: Record<string, any>,
    config?: Record<string, any>,
  ) => Promise<any>;
}